Split routes.py into routes folder with individual py files per blueprint
