+++ backend/templates/dashboard.html
@@ -2,5 +2,5 @@
 {% block content %}
     <h1>Dashboard</h1>
     <p>Welcome, {{ user.name or user.nickname or user.email }}!</p>
-    <p><a href="{{ url_for('logout') }}">Logout</a></p>
+    <p><a href="{{ url_for('auth.logout') }}">Logout</a></p>
 {% endblock %}
