+++ backend/templates/home.html
@@ -1,9 +1,15 @@
 {% extends "layout.html" %}
 {% block content %}
     <h1>Welcome to the Personal Finance Tracker</h1>
-    {% if session %}
+    {# Check the GLOBAL session object for the user key #}
+    {% if session.get('user') %}
         <p>You are logged in!</p>
+        {# 'pretty' variable is still available if passed from render_template #}
+        {% if pretty %}
+        <h2>User Profile Data (from Auth0):</h2>
+        <pre>{{ pretty }}</pre>
+        {% endif %}
     {% else %}
-        <p>Please <a href="{{ url_for('login') }}">log in</a> to continue.</p>
+        <p>Please <a href="{{ url_for('auth.login') }}">log in</a> to continue.</p>
     {% endif %}
 {% endblock %}
