+++ backend/templates/layout.html
@@ -1,25 +1,28 @@
 <!DOCTYPE html>
 <html lang="en">
 <head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Fintrack</title>
-    <style>
-        body { font-family: sans-serif; padding: 20px; }
-        nav a { margin-right: 15px; }
-    </style>
+    
 </head>
 <body>
     <nav>
-        <a href="{{ url_for('home') }}">Home</a>
-        {% if session %}
-            <a href="{{ url_for('dashboard') }}">Dashboard</a>
-            <a href="{{ url_for('logout') }}">Logout</a>
+        <a href="{{ url_for('main.home') }}">Home</a>
+        {% if session.get('user') %}
+            <a href="{{ url_for('main.dashboard') }}">Dashboard</a>
+            {# Display flashed messages #}
+            {% with messages = get_flashed_messages(with_categories=true) %}
+                {% if messages %}
+                    {% for category, message in messages %}
+                        <span class="flash-{{ category }}">{{ message }}</span>
+                    {% endfor %}
+                {% endif %}
+            {% endwith %}
+            <a href="{{ url_for('auth.logout') }}">Logout ({{ session['user'].get('name', session['user'].get('nickname', session['user'].get('email'))) }})</a>
         {% else %}
-            <a href="{{ url_for('login') }}">Login</a>
+            <a href="{{ url_for('auth.login') }}">Login</a>
         {% endif %}
     </nav>
     <hr>
+    {# Add block for flashed messages if not inline #}
     {% block content %}{% endblock %}
 </body>
 </html>
